* moved classes * DoricColumn constructors revised
