+++ src/main/scala/habla/doric/syntax/AllSyntax.scala
@@ -3,7 +3,6 @@ package habla.doric.syntax
 trait AllSyntax
     extends ArrayColumnOps
     with CommonColumnOps
-    with DataFrameOps
     with DStructOps
     with ColumnExtractors
     with LiteralConversions
